New classes and refactorization

+++ src/main/java/online/stworzgrafik/StworzGrafik/shift/ShiftRepository.java
@@ -8,6 +8,6 @@ import java.time.LocalTime;
 import java.util.Optional;
 
 @Repository
-public interface ShiftRepository extends JpaRepository<Shift, Integer>, JpaSpecificationExecutor<Shift> {
+public interface ShiftRepository extends JpaRepository<Shift, Long>, JpaSpecificationExecutor<Shift> {
     boolean existsByStartHourAndEndHour(LocalTime startHour, LocalTime endHour);
 }
